Create Bounty_0_0_3.sol
Contract allows a user, the Maker, to place a bet on a future block being odd.  The user deposits an erc20 token and specifies in how many seconds the bet will settle.  The Maker specifies another address, the Taker, as their opponent.  The Taker can then accept the bet by depositing an equivalent number of identical erc20 tokens.  To close the bet either the Maker or the Taker can call CloseBet and the winner will receive the loser's tokens.

***This code is a proof of concept with numerous vulnerabilities***
1) The mapping structure of bets only allows one bet per erc20 token type per address.  There is no logic to prevent submitting a second bet with a malicious payload that would overwrite a bet already in place.
2) Because of this mapping schema token accounting can be manipulated
3) The bet outcome is pseudo-random based off of a block timestamp which can be manipulated